11104 - refactor apple towards bundleHelper
